Set up project, add reference to DSS, handle dependencies, build
`submods` contains the reference to the whole DSS system in a git repo (keep it up to date).
`runConfigurations` contains the main IDEA run configurations used in the project.
`./pom.xml` defines the modules that need to be compiled with the project and are used by the CLI.
`./dss-cli-main/pom.xml` defines the project for the CLI and the used dependencies from the parent modules.
`SignCLI.java` contains the `main` method for signing documents.